Organize into Packages: Done
Added Task class into ip.duke.task package

Add import codes for use in Duke and Todo classes
