+++ src/main/java/ip/duke/Duke.java
@@ -1,6 +1,7 @@
 package ip.duke;
 
 import ip.duke.exceptions.DukeException;
+import ip.duke.task.Task;
 
 import java.lang.String;
 import java.util.Scanner;
