+++ src/main/java/ip/duke/Todo.java
@@ -1,5 +1,7 @@
 package ip.duke;
 
+import ip.duke.task.Task;
+
 public class Todo extends Task {
     protected Todo(String description) {
         super(description);
